fix from Fedora to better handle 0 byte files

+++ compress.c
@@ -671,13 +671,16 @@ void comprexx(char *fileptr)
 			if (fdout != 1 && close(fdout))
 				write_error();
 
-			if (bytes_in == 0) {
+			if (bytes_in == 0 && force == 0) {
 				if (remove_ofname) {
+					if (!quiet)
+						fprintf(stderr, "No compression -- %s unchanged\n", ifname);
 					if (unlink(ofname)) { /* Remove input file */
 						fprintf(stderr, "\nunlink error (ignored) ");
 						perror(ofname);
 						exit_code = 1;
-					}
+					} else
+						exit_code = 2;
 
 					remove_ofname = 0;
 				}
